0.3.0
updated json data to use ids by exporting a ranId function that can be used by the post route to input a resolved id tag. This granted access to the functionality that if you click a tag it properly displays the note. Will update futhermore to integrate the working usage of the delete function